requirements.txt
pymupdf>=1.23.0
pdfplumber>=0.10.0
groq>=0.9.0
python-dotenv>=1.0.0
streamlit>=1.32.0
reportlab>=4.0.0
pymongo[srv]>=4.6.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0